refactor: extract _axes and _lazy from _state
The batch-axis sublanguage (the `field` chain/data/sample markers and
the `chain_vmap_axes`/`data_vmap_axes`/`trace_sample_axes`/
`get_has_chains` resolvers, plus `chainful_axis`/`chain_to_axis`) and
the `_LazyArray` deferred-construction cluster were buried in the
1800-line `_state.py`, yet they are generic pytree utilities used by
grove, mcmcloop, debug and the BART wrappers. Move them to dedicated
`_axes.py` and `_lazy.py`.

Pure relocation, no logic change. The dependency stays acyclic
(`_lazy` <- `_axes` <- `_state`); `_state` now imports what it needs
from both. `partition_specs`/`_leaf_partition_spec` stay in `_state`
with the other sharding helpers. All importers across src and tests
updated to the new locations.
