+++ src/bartz/debug/_traceconv.py
@@ -37,7 +37,7 @@ from jaxtyping import Array, Float32, UInt
 from bartz._jaxext import minimal_unsigned_dtype
 from bartz.BART._gbart import FloatLike
 from bartz.grove import HeapArrays, TreeHeaps
-from bartz.mcmcstep._state import field
+from bartz.mcmcstep._axes import field
 
 
 def _get_next_line(s: str, i: int) -> tuple[str, int]:
