+++ tests/test_interface.py
@@ -88,7 +88,7 @@ from bartz.mcmcloop import compute_varcount, evaluate_trace
 from bartz.mcmcloop._callback import _tqdm_registry
 from bartz.mcmcloop._loop import _run_mcmc_inner_loop
 from bartz.mcmcstep import State
-from bartz.mcmcstep._state import chain_to_axis, chain_vmap_axes
+from bartz.mcmcstep._axes import chain_to_axis, chain_vmap_axes
 from bartz.prepcovars import GivenSplitsBinner, RangeEvenBinner, UniqueQuantileBinner
 from tests.test_mcmcstep import check_sharding, get_normal_spec, normalize_spec
 from tests.util import (
